Detect table wrapped in print()
If you wrap an R data structure (e.g., vector) in a print() statement,
it will come back from the Jupyter kernel as text only.  We don't have a
clean way to parse or otherwise process those results.  We can detect
that scenario, and for now we will display a message.